+++ backend/node/src/config/env.js
@@ -1,0 +1,3 @@
+import { applyProjectEnv } from "../../../../tools/env-loader.js";
+
+export const serverConfig = applyProjectEnv();